requirements.txt
streamlit
pillow
inference-sdk
